// frontend/src/app/api/models/match.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MatchStatus } from '../models/match-status';
import { Prediction } from '../models/prediction';
import { Score } from '../models/score';
import { Team } from '../models/team';
export interface Match {
  awayTeam: Team;
  group: string;
  homeTeam: Team;
  id: number;
  kickoffAt: string;
  matchday: number;
  myPrediction?: Prediction | null;
  predictionOpen: boolean;
  result?: Score | null;
  status: MatchStatus;
  venue?: string | null;
}